requirements.txt
pandas
scikit-learn
numpy
joblib
pyyaml
matplotlib
pytest 
pytest-mock
tqdm
beartype
rich
pydantic